+++ src/Wizard.js
@@ -211,7 +211,9 @@ export default class Wizard extends Webform {
   }
 
   prepareHeaderSettings(ctx, headerType) {
-    if (this.currentPanel && this.currentPanel.breadcrumb === 'none' || ctx.isSubForm) {
+    const shouldHideBreadcrumbs = this.currentPanel?.breadcrumb === 'none' ||
+      _.get(this.form, 'settings.wizardBreadcrumbsType', '') === 'none';
+    if (shouldHideBreadcrumbs || ctx.isSubForm) {
       return null;
     }
     return this.renderTemplate(headerType, ctx);
@@ -895,26 +897,29 @@ export default class Wizard extends Webform {
   }
 
   setValue(submission, flags = {}, ignoreEstablishment) {
-    this._submission = submission;
-    if (
-      (flags && flags.fromSubmission && (this.options.readOnly || this.editMode) && !this.isHtmlRenderMode()) ||
-      (flags && flags.fromSubmission && (this.prefixComps.length || this.suffixComps.length) && submission._id) ||
-      (this.options.server && (this.prefixComps.length || this.suffixComps.length))
-    ) {
-      this._data = submission.data;
-    }
-
-    if (!ignoreEstablishment) {
-      this.establishPages(submission.data);
-    }
     const changed = this.getPages({ all: true }).reduce((changed, page) => {
       return this.setNestedValue(page, submission.data, flags, changed) || changed;
     }, false);
 
+    if (!flags.sanitize ||
+      (flags && flags.fromSubmission && (this.prefixComps.length || this.suffixComps.length) && submission._id) ||
+      (this.options.server && (this.prefixComps.length || this.suffixComps.length))
+    ) {
+      this.mergeData(this.data, submission.data);
+    }
+
     if (changed) {
       this.pageFieldLogic(this.page);
     }
+
     this.setEditMode(submission);
+
+    submission.data = this.data;
+    this._submission = submission;
+
+    if (!ignoreEstablishment) {
+      this.establishPages(submission.data);
+    }
 
     return changed;
   }