+++ src/addons/index.js
@@ -7,7 +7,11 @@ export const editForms = [
   key: 'settings',
   display: 'form',
   input: true,
-  components,
+  components: components.map((comp) => {
+    comp.tableView = false;
+    return comp;
+  }),
+  tableView: false,
   defaultValue: {
     data: defaultSettings
   },