+++ src/components/columns/editForm/Columns.edit.display.js
@@ -32,6 +32,24 @@ export default [
     ignore: true
   },
   {
+    key: 'hideLabel',
+    ignore: true
+  },
+  {
+    weight: 0,
+    type: 'textfield',
+    input: true,
+    key: 'label',
+    label: 'Label',
+    placeholder: 'Field Label',
+    tooltip: 'The label for this field.',
+    validate: {
+      required: true
+    },
+    autofocus: true,
+    overrideEditForm: true
+  },
+  {
     weight: 150,
     type: 'datagrid',
     input: true,
@@ -94,5 +112,9 @@ export default [
     tooltip: 'Will automatically adjust columns based on if nested components are hidden.',
     key: 'autoAdjust',
     input: true
-  }
+  },
+  {
+    key: 'hideLabel',
+    ignore: true
+  },
 ];