+++ src/components/table/editForm/Table.edit.display.js
@@ -32,6 +32,24 @@ export default [
     ignore: true
   },
   {
+    key: 'hideLabel',
+    ignore: true
+  },
+  {
+    weight: 0,
+    type: 'textfield',
+    input: true,
+    key: 'label',
+    label: 'Label',
+    placeholder: 'Field Label',
+    tooltip: 'The label for this field.',
+    validate: {
+      required: true
+    },
+    autofocus: true,
+    overrideEditForm: true
+  },
+  {
     type: 'number',
     label: 'Number of Rows',
     key: 'numRows',
@@ -105,5 +123,9 @@ export default [
     input: true,
     tooltip: 'Condense the size of the table.',
     weight: 704
+  },
+  {
+    key: 'hideLabel',
+    ignore: true
   },
 ];